requirements.txt
discord.py
yt-dlp
python-dotenv
proxmoxer
requests
urllib3
PyNaCl
spotipy